Move blobChunkAccessorProxy to copy/progress_bars.go
Second part: Move blobChunkAccessorProxy to the newly established
progress_bars.go.

Only moves unchanged code, should not change behavior.
